# perf: Improve execution speed of cleanup_workdir (in dag) 